+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `'offset'` and `'length'` options to cbor_decode().
 - Add `$offset` and `$length` parameters to Cbor\Decoder::add().
 ### Changed
+- '`max_size`' on decode is enforced for indefinite-length elements too.
 ### Removed
 ### Fixed
 - Fix redundant representation of EDN string if text string has some sort of invalid UTF-8 sequences.